agents/requirements.txt
pandas
google-generativeai
openpyxl
dotenv
pyarrow
